Revert "Filter out plansducks by openMRS user"
This reverts commit 17db088a405bbcd76a49d77e40d4aef439dc95a5.
